+++ pheasant/renderers/jupyter/kernel.py
@@ -201,6 +201,8 @@ def output_from_msg(msg: Dict[str, Any]) -> Optional[dict]:
         return dict(type=msg_type, name=content["name"], text=content["text"])
     elif msg_type == "error":
         traceback = "\n".join([strip_ansi(tr) for tr in content["traceback"][1:-1]])
+        if content['ename'] == 'NameError':
+            raise NameError(content['evalue'])
         return dict(
             type=msg_type,
             ename=content["ename"],