+++ tests/renderers/jupyter/test_jupyter_kernel.py
@@ -54,14 +54,10 @@ def test_kernel():
     kernel.execute("a=1")
     assert kernel.execute("a")[0]["data"]["text/plain"] == "1"
     kernel.restart()
-    assert kernel.execute("a")[0]["type"] == "error"
+    with pytest.raises(NameError):
+        kernel.execute("a")
 
     assert kernel is kernels["python"]
-
-    kernel.execute("a=1")
-    assert kernel.execute("a")[0]["data"]["text/plain"] == "1"
-    kernels.restart()
-    assert kernel.execute("a")[0]["type"] == "error"
 
 
 def test_output_hook_factory():